[feature/MOB-245] - 1- Move DaggerWorkerFactory, WorkerSubComponent and WorkerKey from the "support" to the "di" package; 2- Add strings to the notification in SupportNotificationWorker
