+++ app/src/main/java/com/asfoundation/wallet/di/WorkerSubComponent.kt
@@ -1,7 +1,8 @@
-package com.asfoundation.wallet.support
+package com.asfoundation.wallet.di
 
 import androidx.work.RxWorker
 import androidx.work.WorkerParameters
+import com.asfoundation.wallet.support.SupportNotificationWorker
 import dagger.BindsInstance
 import dagger.Subcomponent
 import javax.inject.Provider
@@ -9,6 +10,7 @@ import javax.inject.Provider
 @Subcomponent(modules = [SupportNotificationWorker.Builder::class])
 interface WorkerSubComponent {
   fun workers(): Map<Class<out RxWorker>, Provider<RxWorker>>
+
   @Subcomponent.Builder
   interface Builder {
     @BindsInstance
